Bugfix of issue: Config singleton multiple instantiation issue #4521 (#4523)
