+++ src/CMakeLists.txt
@@ -184,11 +184,13 @@ endif()
 set(valhalla_hdrs
     ${VALHALLA_SOURCE_DIR}/valhalla/valhalla.h
     ${VALHALLA_SOURCE_DIR}/valhalla/worker.h
+    ${VALHALLA_SOURCE_DIR}/valhalla/config.h
     ${VALHALLA_SOURCE_DIR}/valhalla/filesystem.h
     ${VALHALLA_SOURCE_DIR}/valhalla/proto_conversions.h
     )
 
 set(valhalla_src
+    config.cc
     worker.cc
     filesystem.cc
     proto_conversions.cc
